refactor(club-application): Make the completed registration form a component
